edited stop container

+++ scripts/stop_container.sh
@@ -2,5 +2,10 @@
 set -e
 
 # Stop the running container (if any)
-containerid = 'docker ps | awk -F " " '{print $1}''
-docker rm -f $containerid
+containerid=$(docker ps | awk '{print $1}')
+
+if [ ! -z "$containerid" ]; then
+  docker rm -f $containerid
+else
+  echo "No running containers to stop"
+fi
